# Conflicts:
#	package.json
#	resources/js/components/publish/Field.vue

+++ src/GraphQL/Types/AssetInterface.php
@@ -88,10 +88,10 @@ class AssetInterface extends InterfaceType
                 'type' => GraphQL::string(),
             ],
             'height' => [
-                'type' => GraphQL::int(),
+                'type' => GraphQL::float(),
             ],
             'width' => [
-                'type' => GraphQL::int(),
+                'type' => GraphQL::float(),
             ],
             'orientation' => [
                 'type' => GraphQL::string(),
